Added deepfreeze functionallity with password.

+++ com.getshop.core/src/com/thundashop/core/storemanager/StorePool.java
@@ -187,10 +187,14 @@ public class StorePool {
     public synchronized void delete(Store store) throws ErrorException {
         if (store == null) {
             throw new ErrorException(26);
+        } 
+        
+        if (store.isDeepFreezed) {
+            return;
         }
         
-        stores.remove(store.id);
         database.delete(store, credentials);
+        stores.remove(store.id);
         stopStore(store);
     }
